requirements.txt
matplotlib==3.5.2
nextcord==2.0.0a10
Pillow==9.1.1
pony==0.7.16
pyowm-3.3.0